backend/requirements.txt
Flask>=2.2
pymongo>=4.0
python-dotenv>=1.0
Werkzeug>=2.2
pyjwt>=2.8
flask-cors>=4.0
pyserial>=3.5
pyusb>=1.2.1
